refactor: :art: re-organize tests folders
