+++ tests/LSP4-DigitalAsset-Metadata/LSP4Compatibility.spec.ts
