+++ tests/LSP7-DigitalAsset/LSP7.behaviour.ts
