+++ tests/LSP7-DigitalAsset/LSP7.spec.ts
