+++ tests/LSP7-DigitalAsset/extensions/LSP7CappedSupply.spec.ts
