+++ tests/LSP7-DigitalAsset/extensions/LSP7CompatabilityForERC20.spec.ts
