+++ tests/LSP8-IdentifiableDigitalAsset/LSP8.behaviour.ts
