+++ tests/LSP8-IdentifiableDigitalAsset/LSP8.spec.ts
