+++ tests/LSP8-IdentifiableDigitalAsset/extensions/LSP8CappedSupply.behaviour.ts
