+++ tests/LSP8-IdentifiableDigitalAsset/extensions/LSP8CappedSupply.spec.ts
