+++ tests/LSP8-IdentifiableDigitalAsset/extensions/LSP8CompatibilityForERC721.spec.ts
