--- tests/UniversalReceiversProxy.spec.ts
@@ -1,42 +1,0 @@
-import { Signer } from "@ethersproject/abstract-signer";
-import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/signers";
-import { ethers } from "hardhat";
-import {
-  UniversalProfileInit,
-  UniversalProfileInit__factory,
-  UniversalReceiverAddressStoreInit,
-  UniversalReceiverAddressStoreInit__factory,
-} from "../build/types";
-
-const { deployProxy, runtimeCodeTemplate } = require("./utils/proxy");
-
-// const UniReceiver = artifacts.require("BasicUniversalReceiver");
-// const UniversalReceiverTester = artifacts.require("UniversalReceiverTester");
-// const UniversalReceiverAddressStore = artifacts.require("UniversalReceiverAddressStore");
-// const UniversalReceiverAddressStoreInit = artifacts.require("UniversalReceiverAddressStoreInit");
-// const Account = artifacts.require("UniversalProfile");
-
-// keccak256("ERC777TokensRecipient")
-const TOKENS_RECIPIENT_INTERFACE_HASH =
-  "0xb281fc8c12954d22544db45de3159a39272895b169a852b314f9cc762e44c53b";
-// keccak256("LSP1UniversalReceiverDelegate")
-const UNIVERSALRECEIVER_KEY = "0x0cfc51aec37c55a4d0b1a65c6255c4bf2fbdf6277f3cc0730c45b828b6db8b47";
-
-describe.skip("Receivers as Proxies", () => {
-  let accounts: SignerWithAddress[] = [];
-
-  let owner: SignerWithAddress, signer: SignerWithAddress;
-
-  let UniversalProfileMaster: UniversalProfileInit,
-    uniAddressStoreMaster: UniversalReceiverAddressStoreInit,
-    proxyUniversalProfile: UniversalProfileInit,
-    proxyUniAddressStore: UniversalReceiverAddressStoreInit;
-
-  beforeAll(async () => {
-    accounts = await ethers.getSigners();
-    owner = accounts[0];
-    signer = accounts[1];
-  });
-
-  it("test", async () => {});
-});
